+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### next [????.??.??]
+* Make `Data.Tuple.Constraint.ClassNewtype.CTuple0` actually be a class
+  newtype.
+
 ### 0.1.1 [2019.10.21]
 * Split `generator-script` out of the main `.cabal` file, as it is only used
   for development purposes.